Enrure that single message is created per session
A single message is being generated per session window. In order to do
so, the branch DSL is used (to ignore dummy record).

Dummy record is necessary to be sent after the window period of time,
so data can be flushed for output topic as soon as possible, without the
need of waiting till the next round (in this particular case, one round
is being executed per day - and we cannot wait that long).